[!] Fix a couple of bug visible on Linux related to console input: not working console history, special symbols. [!] Fix 'democmd' in simple_plugin always returned VALID value

+++ debugger/src/libdbg64g/services/console/console.h
@@ -59,7 +59,7 @@ private:
     void writeBuffer(const char *buf);
     void processScriptFile();
     bool isData();
-    uint8_t getData();
+    uint32_t getData();
     void clearLine(int num);
     void processCommandLine();
